# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BBYO Connect API
  version: 1.0.0
  summary: Secure closed social network backend for BBYO teens and chapter accounts
  description: |
    Production-oriented backend API for BBYO Connect.

    Security and youth safety features:
    - JWT access tokens + rotating refresh tokens
    - Role-based access control
    - Audit logging
    - Reporting and moderation workflows
    - Language filtering hooks

    Realtime channel (Socket.IO):
    - message.created
    - message.edited
    - message.deleted
    - conversation.updated
    - notification.created
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional error details
    ApiEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          description: Response payload
        meta:
          description: Pagination or extra metadata
        error:
          oneOf:
            - type: "null"
            - $ref: '#/components/schemas/ApiError'
...
    EventCreateRequest:
      type: object
      required: [title, description, startAt, endAt, timezone, visibility]
      properties:
        chapterId:
          type: string
        title:
          type: string
        description:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        timezone:
          type: string
        location:
          type: string
        isVirtual:
          type: boolean
        visibility:
          type: string
          enum: [public, chapter, region, private]
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: List events
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: visibility
          in: query
          required: false
          schema:
            type: string
            enum: [public, chapter, region, private]
      responses:
        '200':
          description: Event page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    post:
      tags: [Events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /events/{id}:
    get:
      tags: [Events]
      summary: Get event by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    patch:
      tags: [Events]
      summary: Update event
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    delete:
      tags: [Events]
      summary: Delete event
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /events/{id}/register:
    post:
      tags: [Events]
      summary: Register for event
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    delete:
      tags: [Events]
      summary: Cancel event registration
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Registration removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /events/{id}/reminders:
    post:
      tags: [Events]
      summary: Create event reminder
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [remindAt]
              properties:
                remindAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /chapters:
    get:
      tags: [Chapters]
      summary: List active chapters
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: regionId
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Chapter page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...
  /chapters/{id}:
    get:
      tags: [Chapters]
      summary: Chapter detail
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Chapter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /chapters/{id}/board:
    get:
      tags: [Chapters]
      summary: Chapter board posts
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Chapter board
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /chapters/{id}/projects:
    get:
      tags: [Chapters]
      summary: Chapter creative projects
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Chapter projects
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /chapters/{id}/travel-intro:
    post:
      tags: [Chapters]
      summary: Create travel intro request
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        '201':
          description: Travel intro created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...